+++ virtio-devices/src/lib.rs
@@ -29,7 +29,6 @@ pub mod epoll_helper;
 mod iommu;
 pub mod mem;
 pub mod net;
-pub mod net_util;
 mod pmem;
 mod rng;
 pub mod seccomp_filters;
@@ -46,7 +45,6 @@ pub use self::epoll_helper::*;
 pub use self::iommu::*;
 pub use self::mem::*;
 pub use self::net::*;
-pub use self::net_util::*;
 pub use self::pmem::*;
 pub use self::rng::*;
 pub use self::vsock::*;
