--- virtio-devices/src/net_util.rs
@@ -1,62 +1,0 @@
-// Copyright (c) 2019 Intel Corporation. All rights reserved.
-//
-// SPDX-License-Identifier: Apache-2.0 AND BSD-3-Clause
-
-use super::{EpollHelper, EpollHelperError, EpollHelperHandler, Queue, EPOLL_HELPER_EVENT_LAST};
-use net_util::CtrlQueue;
-use std::os::unix::io::AsRawFd;
-use std::sync::atomic::AtomicBool;
-use std::sync::{Arc, Barrier};
-use vm_memory::{GuestAddressSpace, GuestMemoryAtomic, GuestMemoryMmap};
-use vmm_sys_util::eventfd::EventFd;
-
-// Event available on the control queue.
-const CTRL_QUEUE_EVENT: u16 = EPOLL_HELPER_EVENT_LAST + 1;
-
-pub struct NetCtrlEpollHandler {
-    pub mem: GuestMemoryAtomic<GuestMemoryMmap>,
-    pub kill_evt: EventFd,
-    pub pause_evt: EventFd,
-    pub ctrl_q: CtrlQueue,
-    pub queue_evt: EventFd,
-    pub queue: Queue,
-}
-
-impl NetCtrlEpollHandler {
-    pub fn run_ctrl(
-        &mut self,
-        paused: Arc<AtomicBool>,
-        paused_sync: Arc<Barrier>,
-    ) -> std::result::Result<(), EpollHelperError> {
-        let mut helper = EpollHelper::new(&self.kill_evt, &self.pause_evt)?;
-        helper.add_event(self.queue_evt.as_raw_fd(), CTRL_QUEUE_EVENT)?;
-        helper.run(paused, paused_sync, self)?;
-
-        Ok(())
-    }
-}
-
-impl EpollHelperHandler for NetCtrlEpollHandler {
-    fn handle_event(&mut self, _helper: &mut EpollHelper, event: &epoll::Event) -> bool {
-        let ev_type = event.data as u16;
-        match ev_type {
-            CTRL_QUEUE_EVENT => {
-                let mem = self.mem.memory();
-                if let Err(e) = self.queue_evt.read() {
-                    error!("failed to get ctl queue event: {:?}", e);
-                    return true;
-                }
-                if let Err(e) = self.ctrl_q.process(&mem, &mut self.queue) {
-                    error!("failed to process ctrl queue: {:?}", e);
-                    return true;
-                }
-            }
-            _ => {
-                error!("Unknown event for virtio-net");
-                return true;
-            }
-        }
-
-        false
-    }
-}
